requirements.txt
discord.py>=1.3.3
redis~=3.5.3
aiohttp~=3.6.2
requests~=2.24.0